requirements.txt
transformers>=4.31.0
trl>=0.7.10
datasets>=2.14.0
peft>=0.4.0
huggingface_hub>=0.16.4
python-dotenv>=1.0.0 
